fix(RoomExtender_PlayAreaGizmo): null play area
The play area gizmo of the RoomExtender looks up the play area in
`Awake` and later uses it when drawing the gizmo. The issue is that the
SDK Manager is not set up correctly at that point in time so the
returned play area is null, resulting in errors when trying to draw the
gizmo.

This fix makes sure `VRTK_SDKManager.instance` is always set up, no
matter when the getter is called. Additionally the play area gizmo
drawing method is changed to not throw any errors if the needed objects
weren't found in the `Awake` method. That method logs a warning already,
so the user is notified about the reason the gizmo isn't drawing.

This fix resolves #985.