feat(SDK): remove force tracking origin on Oculus SDK camera rig
Previously, the OculusSDK CameraRig had the Tracking Origin forced
to floor level inside the SDK. Whilst this is not the default setting
for the OculusSDK and should be the default setting for VRTK, in
forcing it meant that if the opposite settings were required then it
was no longer possible.

In future, the Tracking Origin will need to be set manually to
Floor Level by Oculus SDK users as VRTK will no longer handle it.