fix(Controls): rebuild grabbable slider object to work with snapping
If a connected object is of a different scale to the child slider then
the snapping misbehaves. This is apparently a Unity bug where connected
objects that have a non uniform scale then cause connected joints to
not work correctly.

The solution is to have a dummy object that works as the interactable
object that has a 1,1,1 scale and then scale all children accordingly
to get it to look the same.