chore(Structure): add update prompt editor window
A new prompt window has been added that checks to see if the local
version of VRTK matches the latest release on github. If it does
not match then a prompt window appears informing the user that an
update is available and shows the latest chunk of the changelog.

It also provides a button to open the asset store page for VRTK and
provides an option to hide the popup window if the update isn't wanted.