fix(SDKManager): undo and redo
Undoing or redoing changes to the selected SDKs of the SDK Manager lead
to a broken state of populated object references. The fix is to copy the
selected SDK Info whenever it's changed. That will prevent undos and
redos from changing the SDK Infos the SDK Manager uses as the list of
available SDK Infos. Additionally both the SDK Manager and its Editor
were simplified where possible.