feat(SDKManager): clear scripting define symbols
Whenever one added a vendor SDK to the project the SDK Manager added the
associated scripting define symbols for it. But when one removes the
vendor SDK the symbols can't be removed automatically as there will be
compiler errors after removing the files.

A new button has been added to the SDK Manager editor to clear the used
symbols. This button works even if the project currently has compiler
errors.