feat(SDKManager): disable inspector fields when auto-populating
The SDK Manager can be set up to automatically populate the object
references. The inspector allowed one to change all the referenced
objects, even if the auto-populate checkbox was enabled. This lead to
confusion because the SDK Manager will sometimes override user-made
changes. This change makes sure to disable the object reference fields
that are auto-managed if the SDK Manager is set up to do so. It also has
a tooltip that explains why the fields are disabled and suggests how to
enable them again to customize the object references.