feat(Pointer): add pointer direction indicator prefab
A new prefab has been added that displays a direction indicator on a
Pointer script that is used to denote the desired rotation for the
Destination Marker event.

This can be used now to rotate the play area upon teleporting based
on the direction that the indicator is pointing in.

The Direction Indicator is controller with the touchpad touch angle.

To use the prefab:

 * Add the `PointerDirectionIndicator` prefab to the scene.
 * Select the GameObject with the VRTK_Pointer to add the direction
 indicator to.
 * Drag and drop the `PointerDirectionIndicator` GameObject to the
 VRTK_Pointer `Direction Indicator` inspector parameter.
 * When the scene is running, a direction indicator will follow the
 pointer cursor and the direction the indicator is facing will be
 the direction the play area is facing after teleporting.