feat(Presence): add step height support to body physics
The Body Physics script now supports a step offset height to determine
if a collided object is small enough to be climbed upon even if the
object is too high for the capsule collider to naturally resolve the
collision to the new height.

This is achieved with a sub capsule collider that sits below the main
collider and acts as a foot collider to detect when any objects are
collided with and then a box is cast down to see if the main body
collider should be standing on a new surface.

If the box cast is valid then the play area is either teleported to
the new position if a teleporter is available or the play area is just
snapped to the new position if no teleporter is available.