feat(SDK): allow additional tracked controller usage
The HTC Trackers are now supported in a basic way by allowing for
additional tracked controllers to be supported by the SteamVR SDK
Bridge. The tracker support still requires manual configuration of
the tracker objects and only works with SteamVR. But until more
tracking solutions come out, this will be ok as the HTC trackers
can only be used with SteamVR at the moment anyway.

To set a tracker up:

 * Follow HTC Tracker instructions to set object up with SteamVR
 * Add `VRTK_TrackedController` to tracker object
 * Add `VRTK_ControllerTracker` to tracker object
 * The tracker object is now usable by VRTK
 * Attach any controller scripts directly to the tracker object