feat(Locomotion): add optional location offset for destination point
An optional destination location parameter is now available on the
Destination Point script which allows a custom transform to be
provided which will be used as the destination location for when the
destination point is activated.

This is useful if the destination point needs an offset so the point
that the user ends up at after teleporting isn't directly on the
destination point.