feat(Interaction): allow dependencies to be injected
The Interact Grab and Interact Use script both have a soft dependency
on the Controller Events script and a hard dependency on the
Interact Touch script.

Previously, this dependency was required to be on the same game object
as the script, but now it's possible to inject those dependencies
via an inspector parameter.

The Object Auto Grab script has also been updated to allow for it's
dependencies to be injected.