fix(Logger): saved to scene in Editor
If the Logger was used in the Editor to log something it was added to
scene. This fix makes sure the Logger isn't visible in and is not saved
to the scene when in the Editor.
Additionally the `DontDestroyOnLoad` call has been moved to not be
called when destroying the object anyway.