feat(Pointer): allow bezier pointer to utilise line renderer
The bezier pointer renderer can now utilise a line renderer by
attaching a line renderer to a game object that is provided as the
custom tracer object on the bezier pointer renderer.

If only a line renderer is provided, then only a line renderer will
be used to display the curve of the bezier pointer. If the game object
also contains a mesh the both the mesh and the line renderer is used
in the curve display.