feat(SDK): add methods to get/set boundary draw status
The status of whether a play area boundary is drawn can now be
retrieved and set via the SDK Bridge, which bubbles through to the
relevant SDK to check if the boundaries can be drawn.

At present, only SteamVR allows for checking if the boundary can be
drawn via the `SteamVR_PlayArea.drawInGame` variable.

The SteamVR Boundaries SDK has also been tidied up to cache the
GetComponent call to the SteamVR_PlayArea component.