refactor(Logger): use trace log level by default
When adding the Logger script to the scene to customize its settings the
log level was initially set to `Info` by default, despite the
automatically created Logger using `Trace` instead.
This change makes sure `Trace` is used in all cases by default and
simplifies the automatic creation to use the default values.