chore(gitattributes): allow asset diffs
The `.gitattributes` file was set up to treat various Unity assets as
binary files but they are in fact text files. To allow diff'ing the
files git needs to treat these files as text files. Since treating files
as text files is the default this change removes the `.gitattributes`
file completely, as it doesn't do anything else.