fix(Logger): usage when serialization runs
Whenever the Unity serialization runs the serialization callbacks in SDK
Info are run. Since those methods were updated to use the VRTK Logger
Unity now throws errors because it's not allowed to create new game
objects while Unity is serializing. The fix is to always create an
instance of the VRTK Logger in the Editor.