fix(SnapDropZone): prevent default snap at edit time
The Default Snapped Object parameter would auto snap the given
object in the editor as the SnapDropZone script runs in the editor.

This causes the object hierarchy to change at edit time which is not
a desired outcome.

The fix is to only do the ForceSnap in OnEnable if the scene is at
runtime.