fix(Pointer): defer controller setup to Update method
Previously, the Pointer script would attempt to set up the controller
for the pointer in the OnEnable and then the Start method, however
this caused issues when wanting to use the `activate on enable` setting
along with the `interact with objects` setting as the pointer was
enabled before the controller interactions were created meaning it
would not work.

This fix now ensures the pointer controller set up is done at the first
possible moment in the update routine after all enable and start
methods have been run. This will also mean that the disabling and
enabling of a pointer will still work and set up correctly.