fix(SnapDropZone): clean up on destroy of snapped object
If a snapped object was destroyed (using `Destroy`) then the snap
drop zone would not have it's state reset and the zone would no
longer be usable as it was still in a snapped state.

This fix checks to see if the snapped object is still valid and if
not then it emits the unsnap event.