fix(SnapDropZone): do not require grabbed state on unhighlight
Previously, the snap drop zone would only unhighlight if the
interacting object causing the highlight was being grabbed. This would
cause issues if the object was no longer in the drop zone due to other
reasons such as position changes or joint breaks as the unhighlight
code would not execute as the object wasn't being grabbed.

This would leave the snap drop zone in a permanently highlighted state
so the solution is to ensure the unhighlight doesn't need to consider
the grabbed state to work.