fix(ScreenFade): ensure screen fade reaches full opacity
The VRTK_ScreenFade script could get into a state where it should be
at an opacity of 1f but would end up at 0.9999f which meant the quad
wasn't completely opaque so items in the scene were still barely
visible.

This fix forces the alpha channel to 1f if it exceeds 0.98f.