fix(Interaction): ensure touch rigidbody settings are always applied
The rigidbody settings on the Interact Touch object would sometimes
not be applied if the rigidbody already existed on the object. This
would mean that the colliders of the controller could fall down as
the constraints of the rigidbody were not being fixed.

This fix ensures the settings required are always applied.