chore(Structure): suppress internal obsolete warnings
As a number of core VRTK scripts and methods have been deprecated, it
meant that existing VRTK scripts that still relied on those elements
were using deprecated items.

Usually, when something is deprecated it is replaced with a new item
that can be used in place of the old, however, with some of the
elements that have been deprecated in VRTK, this was not possible.

For example, the `Layers To Ignore` parameter has been replaced with
a more complex Custom Raycast script so a direct replacement is not
possible as it is a paradigm change in how things work. The original
`Layers To Ignore` cannot simply be removed from the internal scripts
as this would break existing behaviour (and therefore not be a valid
deprecation).

This meant that deprecation warning messages were being output to
the console and as there were many of these internal deprecation
cases, the console was being filled with warning messages, which
was annoying and unsightly.

The fix is to wrap any internal use of an obsolete item in the
`#pragma warning disable 0618` tag to ensure any internal deprecation
messages are suppressed.

When the deprecated items are finally removed, these additional
tags will also be removed as the offending lines will no longer raise
any deprecation warning methods.