fix(DeviceFinder): headset type recognition
The headset type recognition has a fixed list of known headsets.
This list is missing a special Vive Pre model that is reported as
"Vive DVT".
This fix adds that headset and modifies the recognition to also tell
the user to report any new headset type so it can be added to VRTK.
Additionally a slower check is done when a summarization is
requested to support types VRTK doesn't know about yet.

This fixes #1113.