feat(Utilities): add sdk object alias script for following SDK objects
The SDK Object Alias script will child the script's GameObject to the
selected SDK Object.

This is useful if a GameObject needs to follow the SDK headset or
play area as these can change during runtime if the SDK is switched.

All of the example scenes that used the Transform Follow have been
updated now to use this new alias script to know which object to
follow as the SDK Object is likely to change when switching SDK.