fix(Tooltips): initialise tips when controller is enabled
The Controller Tooltips would attempt to initialise the tips every
frame in an Update loop until they were correctly initialised.

This would cause an issue if the controller elements were not available
as the Initialise method would run every Update.

The fix now only initialises the tips when the controller enabled event
is thrown which is a more appropriate time to set the tooltips up.

Another fix was added that prevented the tooltips from being visible
if the `Hide when not in view` option was unticked.