feat(Controller): add audio haptics
The audio haptics vibrate the controller based on an AudioClip
which uses the audio clip data for the pattern of the vibration.

This is useful for providing more customised vibrations.

The InteractHaptics script has been updated now to accept an
audio clip as a parameter for touch/grab/use and comes with a
custom editor to hide the default settings if a clip is provided.

An additional fix has also been added for the Oculus SDK which
was preventing the default haptic pulse from working due to
the way the OVRHapticsClip was being set up before the Oculus
SDK was available.