fix(Interaction): prevent secondary grab when isGrabbable is false
There was an issue that if a secondary grab mechanic was set then it
was still possible to grab an object even if the `isGrabbable` option
was set to false.

This was due to the `PerformSecondaryAction` method would not take
into consideration if a primary grab had already occurred. The fix
is to check if a primary grab has occurred in that method.