feat(Version): add versioned symbols for VRTK
A new script has been added that defines the current and previous
VRTK versions and adds a scripting define symbol to the player
settings for it. This symbol is useful for third party assets that
use VRTK.

There's an exact version symbol for the current version, as well as
additional `_OR_NEWER` symbols for the current and any previously
known versions of VRTK. They follow this pattern:

- `VRTK_VERSION_3_2_0`
- `VRTK_VERSION_3_2_0_OR_NEWER`

Additionally the Support Info editor window was updated to list this
VRTK version. A note hints that this version may not be correct in
case the GitHub version of VRTK is used. This change also makes sure
the Support Info window has the correct icon.