fix(Presence): prevent crashes when switching scenes with body physics
There is an issue where if an array is not initialised by the time
the scene has loaded (because Unity initialises arrays in the
inspector when the scene is played) it would cause Unity to crash.

This meant that any scene that used Body Physics would crash if the
scene was attempted to be loaded via the SceneManager because Unity
does not initialise the array unless it's starting from that scene
in the editor.

There was another issue with the Headset Collision script randomly
crashing due to a race condition and the script no longer being
available.