feat(Structure): add events for all current actions
A number of the scripts perform actions that set state but the only
way of knowing the action has happened or the state has changed was
to have a reference to the script and access relevant variables or
methods.

This change sees that any VRTK script that performs an action will
now also emit a C# delegate event (and a corresponding Unity event
via the Unity Event helpers). This means that any action that occurs
can be utilised simply through registering for the event.