chore(Examples): update scenes to utilise new SDK Transform Modify
The example scenes have a collection of grabbable objects that use
snap handles, however these snap handles are all set to the orientation
of the HTC Vive wand in SteamVR and therefore objects don't grab at
an appropriate orientation when using another controller such as the
Oculus Touch.

All of the relevant scenes have now been updated to ensure they
use the SDK Transform Modify.

Plus all scenes have had a bit of a tidy in the hierarchy so they
are neater for others to understand.