fix(Examples): prevent spawning multiple arrows when holding bow
There was an issue where the arrow spawner would spawn multiple arrows
when holding the bow because the Oculus Avatar will child the bow
object to the model alias and not the actual controller game object.

The ArrowSpawner code would only check the model alias and not the
actual game object. This fix ensures both are checked.