feat(Pointers): expose policy list as public on destination marker
The DestinationMarker script had a hidden policy list that the
teleporter would use to determine when to set a pointer to the valid
or invalid colour. However, it wasn't possible to just change the
pointer state based on simply what the pointer was looking at.

This change exposes the hidden policy list variable on the destination
marker and makes it a public parameter that can be updated in the
Unity inspector.

If the pointer has a policy list assigned then it won't get assigned
one by the teleporter, this is useful for overriding the teleport
policy list setting.