feat(Pointer): emit event when pointer state is valid or invalid
The Pointer script now emits an event when the pointer state changes
between being valid and invalid, which is effectively the same as when
the pointer changes between the valid colour and the invalid colour.

The event is emitted with a destination marker payload containing the
RaycastHit data of the object currently being colliding with.

There is also an additional public method that returns the state
validity of the pointer in case an event isn't required.