fix(SDKManager): SDK Setup load/unload crashes on scene changes
Whenever the scene changed it was recommended to unload the
currently loaded SDK Setup via the SDK Manager. This lead to issues
with coming up with the exact time to call the unload method,
especially when using `SteamVR_LoadLevel`. That SteamVR script even
lead to crashes.

This fix stops unloading the currently used VR Device everywhere.
The result is that **there is no need to call the unload method
manually anymore**.

The SceneChanger script was simplified a bit after being updated for
this fix.