fix(Utilities): new vive keyword for headset discovery in device finder
The DeviceFinder script has a method for determining which headset is
being used. However, there seemed to be an issue where newer HTC Vives
would report their name as `Vive MV.` with a dot at the end rather than
the standard `Vive MV` which was reported by the original HTC Vive
consumer edition.

This fix just checks both values and assumes the headset is `ViveMV`
to prevent users with newer Vive headsets receiving a warning.