feat(Utilities): simplify headset checking in device finder
The `GetHeadsetType` method in the DeviceFinder script has now been
simplified to parse the current headset string and process it by
removing any spaces and full stops `.` so a pure string can be
compared.

This processing of the VR device string is then cached so further
processing is not required.

A new method has been added that resets the headset cache if for
some reason the headset type changes at runtime then the cache can
be reset so it is re-parsed and re-processed.