fix(Locomotion): prevent current destination point being set to null
There was an issue with the DestinationPoint script where the
`currentDestinationPoint` which is a static variable was being reset
to null whenever the destination point was set or the pointer exited
the destination point area.

This was not apparent when the pointer activation button and selection
button was the same button because of the speed in which the events
would be triggered.

However, if the activation and selection buttons were different then
the `currentDestinationPoint` would be set to null and this would
reactivate the actual current destination point even though it should
be disabled.

The fix is to ensure the `currentDestinationPoint` is never set to
null and is used for a simple check when resetting the Destination
Point.