fix(SDK): provide alternative way of getting unknown assembly types
Previously, the SDK Defines would require the use of an assembly name
when attempting to get a type and if it was not found then it would
cause an error.

This fix adds a new method that attempts to get an unknown assembly
type and if it can't be found then it returns `null`.