fix(SDK): SteamVR SDK Setup unusable in build
The SteamVR scripting define symbols were defined based on
`SteamVR_Update.currentVersion`, but that type is only available in
the Editor assembly, which is not included in builds. The fix is to
revert back to the old method of identifying the used SteamVR plugin
version and to add new identification logic for the SteamVR plugin
version 1.2.2.