fix(SDK): ensure simulator position is correct
  > **Breaking Changes**

The example scenes were using a script to position the simulator rig
correctly. This is no longer needed after recent changes to the
simulator and therefore the script and its usage in the `SDKSetups`
prefab is removed.

If users of VRTK used this script previously or broke the prefab
link to the `SDKSetups` prefab they need to remove any usage of the
script.