fix(SnapDropZone): ensure force snap object state is saved correctly
There was an issue with the Snap Drop Zone that if on scene load an
interactable object was to be snapped to it, the order of load would
mean that the interactable object would not have completed set up
and therefore would not save the state of the interactable object
on force snap.

The fix is to ensure all force snaps happen at end of frame, removing
any race conditions with script load order.