refactor(Interaction): apply best practice coding standards
The cases where an interactable object checks for a null value have
been updated to ensure they do an explicit check. Also the foreach
loops have been replaced with for loops as they are more efficient
in Unity 5.5 and below.

The Object Auto Grab script also now correctly cancels the
coroutine if the script is disabled.