fix(RadialMenu): disable mouse interactions
The radial menu prefab was using a `GraphicRaycaster` which meant that
the canvas used for the menu was responding to the mouse. The fix is to
either add a `VRTK_UICanvas` to the prefab in case one wants to interact
with the menu via the VRTK pointers or to remove the raycaster. This fix
chooses the latter as users of VRTK probably only want to enable usage
of the radial menu by touchpad only.