fix(RadialMenu): buttons have colored background
The radial menu buttons all had a colored background added to them
which is white by default. The issue was that a previous commit
removed the needed `OnPopulateMesh` method because that one was
using a deprecated argument instead of the one it was replaced with.
Instead of removing said method this fix makes sure to change the
method to take the new argument.