feat(SDK): implement internal Unity VR support
Unity supports basic VR functionality without the need for any
third party SDK. A new SDK bridge has been added that hooks into
this internal Unity VR support so VRTK can be used for any device
that Unity supports.

The InputManager has been update to have the default axis information
required for the axes on the controllers to work correctly.

A new VelocityEstimator script has been added that enables better
estimations of velocities of moving objects that are not rigidbodies.
This is used to determine the velocity and angular velocity of the
HMD and controllers when using the Unity VR SDK.

The Velocity Estimator is largely based on the velocity estimator
script from Valve's SteamVR Interaction System, so big thanks
to Valve for this solution!