feat(SDKManager): simplify single SDK Setup configurations
Supporting multiple SDKs means VRTK had to disable all the SDK
Setups that are used automatically to ensure they're usable for
runtime switching. This change allows to completely bypass the
runtime loading and that requirement in case only one SDK Setup is
used. This means users of VRTK don't have to bother with handling
when an SDK Setup is loaded and usable, nor do they have to
re-enable their SDK Setups game objects again to work on them in the
editor while they're visible.

The SteamVR SDK Setup in the SDKSetups prefab is changed to allow
for this new behavior. This also fixes a bug that resulted in
invalid matrix errors being logged to the console by Unity.