refactor(SnapDropZone): reduce GetComponent calls in update
The SnapDropZone script would continually call `GetComponent` in the
`Update` method as well as in the `OnTriggerStay` method to attempt
to get the Interactable Object component of the snappable GameObject.

This was an inefficient call and has been replaced with a cached
version of the interactable object to ensure the inefficient
`GetComponent` call is removed from the `Update` and `OnTriggerStay`
methods.