feat(Pointer): allow direction indicator to be visible with cursor
A new option has been added to the Pointer Direction Indicator
so it can always be visible when the pointer cursor is visible.

This allows for the `Always On` setting for the cursor on the
pointer renderer to always determine if the direction indicator
is also on.

Even if the `Always Off` setting is on for the cursor then the
pointer direction indicator will still become visible when
the pointer is activated.