fix(Examples): prevent crash on scene change with active arrow
When the arrow prefab had collided with something it would re-enable
the interactable object on the arrow causing a chain of events that
could cause a crash if the scene was being changed at the point
this was initiated.

This fix checks to make sure the arrow is still active before trying
to reset it.