fix(Utilities): prevent crash when switching sdk setup
There was an issue with the `LoadedSetupChanged` method that it
would attempt to access objects or destroy objects when the SDK
was finally unloaded because the object in question had already
been destroyed (e.g. when ending the scene).

This fix just ensures the sdk setup changes only call their intended
methods if the game object the script is on is still active in the
hierarchy.