refactor(Structure): replace occurrences of Equals with ==
Unity3d creates garbage when dealing with `Equals` as it's considered
a Boxing method and therefore is less efficient than simply comparing
with `==`

This information can be found at: goo.gl/apMyGU

This refactor replaces all of the occurrences of `Equals` with `==`
to provide the efficiency improvements.

The MoveInPlace script has also been slightly refactored to use a
`switch` statement rather than a collection of `if/else` statements
that all in turn used the `Equals` method as well.