fix(Utilities): use local space for velocity estimation samples
The Velocity Estimator was using world space to determine the samples
for velocity and angular velocity, which would cause problems if the
Velocity Estimator was attached to a GameObject that's parent was
rotated.

It could be seen in the Unity SDK by simply rotating the CameraRig
around the Y axis and then the throwing mechanism would throw
things in the wrong direction.

The solution is to simply use local space for position and rotation
as then the actual object the Velocity Estimator is attached to
will be used for the relative position and rotation.