chore(SDK): replace simulator velocity calculations
The Simulator SDK has had the velocity calculations replaced to use
the new Velocity Estimator script as it provides a better estimated
velocity and means the logic is done in one concise place.

The Simulator CameraRig prefab has been updated to contain the scripts
required. But if an older version of the CameraRig prefab is being
used then the scripts will automatically be added by the SDK.