fix(Interaction): prevent coroutine running if game object is disabled
The Interact Object Appearance script was throwing an error if the
co routine to determine the appearance was being run but the script
GameObject had been disabled (e.g. when the play mode was stopped).

This has been fixed by checking that the GameObject is active in
the hierarchy to determine if it's a valid interaction.