fix(Interaction): determine grab rotation based on quaternion not eular
  > **Breaking Changes**

The Grab Attach Mechanics were using a Eular angle to determine the
grabbed rotation which meant the actual rotation values were odd even
though the grabbed rotation was correct.

The Track Object grab mechanic didn't use a Eular and used an
inverted Quaternion which was correct but required the snap handles
to be flipped to work as the other grab mechanics worked with the
snap handles.

As the other grab mechanics now all use the quaternion and not the
eular to determine their rotation, it means that all snap handles
work the same and don't require any flipping.

This also means that the existing values for any snap handles will
be incorrect and will need inverting to work correctly again.

All of the example scenes have been updated to reflect the breaking
change and all of the snap handles have been reconfigured to work
correctly.