fix(Interaction): remove valid controller check from secondary action
The Secondary Grab Action was checking to see if the controller was a
valid grab controller on the interactable object.

This did not make sense as if the right controller was a valid grab
controller then only the right controller could pick up the object
but then the left controller could never initiate a secondary grab
action as the right controller was the only one denoted as being
valid to initiate a grab action.