fix(Interaction): ensure controllers preserve the correct visibility
The Interact Object Appearance script was incorrectly hiding controller
visibility when both controllers would touch the object.

This was because the cached object state was only for the last object
that interacted, however an existing interacting controller could
be overridden by a new interacting controller which would then prevent
the previous controller from regaining the correct visibility state.

The fix is to ensure the object state is cached accordingly.

Another issue was how it was storing the coroutines for each
interaction event as it was storing them per interaction not per
object that was interacting. This would mean that if both controllers
touched the same object at the same time, then one would cancel the
coroutine of the other. This has been fixed by grouping the
coroutines with the object they are to affect.