fix(RadialMenu): ensure button placement is calculated correctly
The angles of the generated buttons was being calculated incorrectly
due to a division by an integer rather than a float which caused
certain button amounts to unevenly distribute around the radial
menu (e.g. 7 buttons).

Any reference to a float should use a float in the calculation or
equation by defining the `f` suffix to ensure it doesn't fall back
to using integers.

Thanks to @rivierasolutions for the fix.